% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bayesLm.R
\name{bayesLm}
\alias{bayesLm}
\title{MCMC algorithm for Bayesian Linear Model}
\usage{
bayesLm(y, X, mu0, Sigma0, a0, b0, sigma2Int, nkeep = 10000, nburn = 1000)
}
\arguments{
\item{y}{vector of repsonse}

\item{X}{design matrix for regression. If intercept is desired, a column of 1's is needed}

\item{mu0}{prior mean for beta}

\item{Sigma0}{prior var-cov matrix for \eqn{\beta}}

\item{a0}{prior shape parameter for \eqn{\sigma^2}}

\item{nkeep}{number of iterations to keep}

\item{nburn}{number of iterations to toss}

\item{a0, b0}{prior scale parameter for \eqn{\sigma^2}}

\item{sigma2Init}{initial value for \eqn{\sigma^2} in MCMC}
}
\value{
list with mcmc sample and mean fitted values
}
\description{
Fits the linear regression model
\eqn{\beta~N(\mu_0, \Sigma_0)},
\eqn{\sigma^2~IG(a0, b0)},
\eqn{y~N(X\beta, \sigma^2 I)}
}
\details{
Uses Gibbs sampling to sample from the posterior under the above linear regression model.
}
