% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_TdistLm.R
\name{bayesTdistLm}
\alias{bayesTdistLm}
\title{MCMC algorithm for Heavy Tailed (t-distribution) Bayesian Linear Model}
\usage{
bayesTdistLm(y, X, mu0, Sigma0, a0, b0, parInit = NULL, nu, nkeep = 10000,
  nburn = 1000, rwTune = NULL)
}
\arguments{
\item{y}{vector of repsonse}

\item{X}{design matrix for regression. If intercept is desired, a column of 1's is needed}

\item{mu0}{prior mean for beta}

\item{Sigma0}{prior var-cov matrix for \eqn{\beta}}

\item{a0}{prior shape parameter for \eqn{\sigma^2}}

\item{b0}{prior scale parameter for \eqn{\sigma^2}}

\item{parInit}{vector of initial values for \eqn{\beta, \sigma^2}. Default is \code{NULL} in which case initial values are chosen using a robust regression using the default settings in \code{\link[MASS]{rlm}}}

\item{nu}{degrees of freedom for t-distribution - assumed fixed}

\item{nkeep}{number of iterations to keep}

\item{nburn}{number of iterations to toss}

\item{rwTune}{tuning parameter for the Metropolis-Hastings random walk used to sample \eqn{\sigma^2}}
}
\value{
list with mcmc sample, mean fitted values, and logical acceptance vector for \eqn{\sigma^2} proposals.
}
\description{
Fits the heavy-tailed linear regression model
\eqn{y_i~N(x_i'beta,v_i)},
\eqn{v_i~inv-chi^2(nu,sigma2)}
\eqn{\beta~N(mu0,Sigma_0)}
\eqn{sigma2~IG(a0,b0)}.
}
\details{
Uses Gibbs sampling to sample from the posterior under the above linear regression model. This model is equivalent to a linear regression with t-distributed errors.
}

