% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{beta_ls}
\alias{beta_ls}
\alias{fn.sample.beta.l}
\alias{fn.one.rep.beta.l}
\title{Sampling for beta_l's in hierarchical model
5. [beta_l|-] l=1,..,nGroups}
\usage{
fn.sample.beta.l(yl, Xl, XtX.l, Beta, sigma2l, bstar, Sigma0Inv)

fn.one.rep.beta.l(y, X, XtX, Beta, sigma2, bstar, betalMat, Sigma0Inv)
}
\arguments{
\item{yl, Xl}{these are the responses and design matrix for the lth group}

\item{XtX.l}{X'X for lth group}

\item{Beta}{Beta}

\item{bstar}{b=b^*/swSq}

\item{y, X}{list of yl and Xl}

\item{XtX}{list of X'X for the l groups}

\item{sigma2l:}{the sigma^2_l for the l group}
}
\description{
Sampling for beta_l's in hierarchical model
5. [beta_l|-] l=1,..,nGroups
}
\details{
Sigma0Inv: fixed  loop through for each group and do sampling for each group beta
}
