% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_heirarchical_tmodel.R
\name{fn.sample.beta.lTmodel}
\alias{fn.sample.beta.lTmodel}
\alias{fn.sample.v.lTmodel}
\alias{fn.one.rep.betasAndVs}
\title{Sampling for beta_l's and v_ls in hierarchical T-model}
\usage{
fn.sample.beta.lTmodel(yl, Xl, SigmaVlInv, Beta, bstar)

fn.sample.v.lTmodel(yl, Xl, betal, sigma2l)

fn.one.rep.betasAndVs(y, X, Beta, sigma2, betalMat, vlList, bstar)
}
\arguments{
\item{yl, Xl}{desc}

\item{SigmaVlInv}{desc}

\item{Beta}{desc}

\item{bstar}{desc}

\item{betal}{desc}
}
\description{
Sampling for beta_l's and v_ls in hierarchical T-model
}
\details{
Sampling beta.l and v.l for the T-model. fn.one.rep.betasAndVs puts the two together to sample both for all groups.
}
