% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brlm.R
\docType{package}
\name{brlm}
\alias{brlm}
\alias{brlm-package}
\title{brlm: A package for implementing Bayesian Restricted Likelihood Methods.}
\description{
brlm: A package for implementing Bayesian Restricted Likelihood Methods.
}

