% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{bstar_step}
\alias{bstar_step}
\alias{fn.loglike.bstar}
\alias{fn.proposal.logbstar}
\alias{fn.logpdf.prop.logbstar}
\alias{fn.sample.bstar}
\title{functions for bstar step in hierarchical model}
\usage{
fn.loglike.bstar(bstar, v1, v2, quad1, quad2, K, p)

fn.proposal.logbstar(logbstarCur, step_logbstar)

fn.logpdf.prop.logbstar(bstarProp)

fn.sample.bstar(bstarCur, v1, v2, quad1, quad2, K, p, step_logbstar)
}
\arguments{
\item{bstar}{bstar}

\item{v1}{v1}

\item{v2}{v2}

\item{quad1}{quad1}

\item{quad2}{quad2}

\item{K}{K}

\item{p}{p}

\item{logbstarCur}{description}

\item{step_logbstar}{description}

\item{bstarProp}{description}

\item{bstarCur}{description}
}
\description{
functions for bstar step in hierarchical model
}
\details{
functions for bstar step
}
