% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixture.R
\name{fitMixture}
\alias{fitMixture}
\title{A standard Bayesian mixture model for outliers using a mixture of two normal distributions for the likelihood.}
\usage{
fitMixture(y, theta.lims, length.theta, sigma2.lims, length.sigma2, mu, tau,
  alpha, beta, p, c)
}
\arguments{
\item{y}{vector of data}

\item{tau}{mean and standard deviation of the normal prior distribution on \eqn{\mu}}

\item{alpha}{prior shape and scale of inverse gamma prior on \eqn{\sigma^2}}

\item{beta}{prior shape and scale of inverse gamma prior on \eqn{\sigma^2}}

\item{p, c}{fixed values for p and c in above model}
}
\value{
A list of length 4: the joint posterior, the two marginals, and the posterior means
}
\description{
Fiting the following mixture model: \deqn{
\theta\sim N(\mu, \tau^2),
\sigma^2 \sim IG(\alpha, \beta),
y_i~(1-p)N(\theta, \sigma^2)+pN(\theta, c*\sigma^2)
}
where p and c are fixed by the user
}
\details{
This implementation is fit on a grid of \eqn{\theta} and \eqn{\sigma^2} using Riemann sum numerical integration to find the marginal distribution of the data in a similar way as in \code{\link{fitOrderStat}}. This is rather naive but is relatively quick - it only works well if the choice of grid on which the Riemann sum is done is chosen carefully. The grid is specified by \code{theta.lims}, \code{sigma2.lims}, \code{length.theta}, and \code{length.sigma2}. The grid should cover the region of non-negligble posterior mass. The number of grid points (\code{length.theta}/\code{length.sigma2}) must be large enough for good precision. However, larger values increase computation time.
}
\seealso{
\code{\link{fitOrderStat}}
}
