% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOrderStat.R
\name{fitOrderStat}
\alias{fitOrderStat}
\title{Restricted likelihood Bayesian model using a middle set of order statistics to define the restricted likelihood.}
\usage{
fitOrderStat(y, k, mu_lims, length_mu, sigma2_lims, length_sigma2, eta, tau,
  alpha, beta)
}
\arguments{
\item{y}{vector of data}

\item{k}{numeric: choice of k determining the set of order statistics on which to condition.}

\item{mu_lims, sigma2_lims}{vectors of length 2 specifying the lower and upper limits for the numerical integration. Should span region of non-negligble posterior probability, otherwise the posterior will be incorrect}

\item{length_mu, length_sigma2}{Number of grid points for each parameter used in the numerical integration}

\item{eta, tau}{mean and standard deviation of the normal prior distribution on \eqn{\mu}}

\item{alpha, beta}{prior shape and scale of inverse gamma prior on \eqn{\sigma^2}}
}
\value{
A list of length 4: the joint posterior, the two marginals, and the posterior means
}
\description{
Conditioning on a set of order statistics under a Guassian location and scale model.
The full model is: \deqn{
\mu\sim N(\eta, \tau^2),
\sigma^2 \sim IG(\alpha, \beta),
y_i~N(\mu, \sigma^2)
}
Conditioning is on a middle set of order statistics \eqn{(Y_{(k+1)}, \dots Y_{(n-k)})}. That is, the likelihood used in the Bayesian update is the likelihood of these order statistics. When \eqn{k=0}, the model is the full normal theory model written above.
}
\details{
Fit is done a grid of \eqn{\mu} and \eqn{\sigma^2} using Riemann sum numerical integration to find the marginal distribution of the data. This is rather naive but is relatively quick - it only works well if the choice of grid on which the Riemann sum is done is chosen carefully. The grid is specified by \code{mu_lims}, \code{sigma2_lims}, \code{length_mu}, and \code{length_sigma2}. The grid should cover the region of non-negligble posterior mass. The number of grid points (\code{length_mu}/\code{length_sigma2}) etast be large enough for good precision. However, larger values increase computation time.
}
\seealso{
\code{\link{fitMixture}}
}
