% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_compute_ystst.R
\name{fn.comp.ystst}
\alias{fn.comp.ystst}
\title{Compute y**
Computes y** from a proposed y* value}
\usage{
fn.comp.ystst(
  yst,
  X,
  l1obs,
  s1obs,
  psi = psi.huber,
  scale.est = "Huber",
  maxit = 400
)
}
\arguments{
\item{yst}{data vector to be transformed to ystst satisfying the observed summary statistics}

\item{X}{design matrix}

\item{l1obs, }{s1obs: observed statsitics}

\item{maxit}{number of iterations to find T(yst)}

\item{psi:}{psi function returning psi(x)/x or psi'(x). default is psi.huber from \link{MASS}}

\item{scale.est:}{defines the scale estimator. Currently supports only 'Huber'}
}
\value{
ystst a data vector such that T(ystst)=T(y_obs)
}
\description{
Compute y**
Computes y** from a proposed y* value
}
