% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_grads.R
\name{fn.grads}
\alias{fn.grads}
\title{Computes the gradients in the linear regression MCMC}
\usage{
fn.grads(y, X, l1obs, s1obs, fn.psi = fn.psi.huber, fn.chi = fn.chi.prop2)
}
\arguments{
\item{y}{numeric vector of responses}

\item{X}{design Matrix}

\item{l1obs, s1obs}{observed value of slope and scale statistics}

\item{fn.psi}{psi function for M-estimation}

\item{fn.chi}{chi function for M-estimation}
}
\description{
Computes the gradients in the linear regression MCMC
}
