% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_one_rep_y.R
\name{fn.one.rep.y}
\alias{fn.one.rep.y}
\alias{fn.one.rep.y2}
\title{Perform a single Metropolis-Hastings step}
\usage{
fn.one.rep.y(y.curr, Beta, sig, l1obs, s1obs, X, log.prop.den.curr, proj,
  fn.psi, fn.chi, psi, scaleEst, maxit = 400)

fn.one.rep.y2(y.curr, Beta, sig, l1obs, s1obs, X, log.prop.den.curr, proj, Qt,
  fn.psi, fn.chi, psi, scaleEst, maxit = 400)
}
\arguments{
\item{y.curr}{current data vector}

\item{Beta}{vector of regression coefficient parameters}

\item{sig}{sigma-the standard deviation parameter}

\item{l1obs}{conditioning statistic-estimate of Beta from observed data}

\item{s1obs}{conditioning statistic-estimate of sigma from observed data}

\item{X}{regression design matrix (i)}

\item{log.prop.den.curr}{value of the proposal density for the current data vector (free of other parameters so it can be saved from iteration to iteration to save computation time)}

\item{proj}{the projection matrix onto the deviation space (aka the least squares residual space)}

\item{fn.psi}{custom psi function returning psi(x) or psi'(x) defining M-estimator for location ; default is fn.psi.huber}

\item{fn.chi}{custom chi function returning chi(x) or chi'(x) defining M-estimator for scale; default (and currently only capability) is fn.chi.prop2}

\item{psi}{psi function returning psi(x)/x. Must match fn.spi, but currently no warning is thrown.}

\item{scaleEst}{'Huber' (default and currently only capability).}

\item{maxit}{max iterations for the estimation of the summary statistics}

\item{Qt}{Q transpose where Q is the orthonormalized X}
}
\value{
list y.curr: the new (if accepted) or same (if rejected) data vector, a: indicator of acceptance of proposed value 1 if accepted, 0 if rejected, log.prop.den.curr: proposal density on log scale of the returned data vector, rat: MH ratio, log.rat1: log ratio of proposal densities
}
\description{
Performs a single Metropolis-Hastings step for the restricted likelihood method
}
\details{
Two equivalent versions, the second uses \code{fn.attenuation2} and is faster. This function is designed for use within an MCMC sampler. Many of the inputs are iteration dependent. The input \code{log.prop.den.curr} is designed as a time saver. Since this is free of other parameters it can be saved from iteration to iteration to save computation time. Note the use of both fn.psi (custom in the \link{brlm} package) and psi (from the \link{MASS} package). This is poor coding and should be corrected.
}
