% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_radius.R
\name{fn.radius}
\alias{fn.radius}
\alias{fn.radius2}
\title{Computes radius}
\usage{
fn.radius(ystst, X)

fn.radius2(ystst, proj)
}
\arguments{
\item{ystst}{a data vector such that T(ystst)=T(y_obs)}

\item{X}{design matrix}

\item{proj}{the projection matrix onto the deviation space (aka the least squares residual space)}
}
\value{
The radius in deviation space
}
\description{
Compute the radius in deviation space
}
\details{
Two equivalent ways to calculate radius in deviation space. The second uses the  projection matrix onto the deviation space as an argument is a bit quicker.
}

