% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{fn.sample.Beta}
\alias{fn.sample.Beta}
\title{sampling for beta in hieararchical model
[Beta|-]}
\usage{
fn.sample.Beta(betalMat, bstar, nGroups)
}
\arguments{
\item{betalMat, }{matrix with columns equal to current beta_l's from each group}

\item{bstar:}{current value of bstar}
}
\description{
sampling for beta in hieararchical model
[Beta|-]
}
\details{
Sigma0 is fixed,  nGroups (number of groups) and swSq are fixed, and used within this function.
}
