% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_heirarchical.R
\name{fn.compute.ab}
\alias{fn.compute.ab}
\title{sampling mu in hierarchical model of paper.
[mu_bstr|-]
function to compute shape parameters of beta from the mean mu<-a/(a+b) and psi=a+b (precision)}
\usage{
fn.compute.ab(mu, psi)
}
\arguments{
\item{mu}{= a/(a+b)}

\item{psi}{= a + b}
}
\description{
sampling mu in hierarchical model of paper.
[mu_bstr|-]
function to compute shape parameters of beta from the mean mu<-a/(a+b) and psi=a+b (precision)
}
\details{
a, b are the shape parameters of a beta distribution
}
