% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{mu_rho}
\alias{mu_rho}
\alias{fn.loglik.mu_rho}
\alias{fn.proposal.mu_rho}
\alias{fn.sample.mu_rho}
\title{Sampling mu_rho
[mu_rho|-]}
\usage{
fn.loglik.mu_rho(mu_rho, psi_rho, rho)

fn.proposal.mu_rho(mu_rhoCur, mu_rho_step)

fn.sample.mu_rho(mu_rhoCur, psi_rho, rho, mu_rho_step)
}
\arguments{
\item{mu_rho, }{mu_rhoCur}

\item{psi_rho}{descrip}

\item{rho}{descrip}

\item{mu_rho_step}{descrip}
}
\description{
Sampling mu_rho
[mu_rho|-]
}
\details{
sampling mu_rho in hierarchical model
}
