% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_functions.R
\name{psi}
\alias{psi}
\alias{fn.psi.huber}
\alias{fn.psi.bisquare}
\alias{fn.chi.prop2}
\title{Psi functions}
\usage{
fn.psi.huber(u, k = 1.345, deriv = 0)

fn.psi.bisquare(u, c = 4.685, deriv = 0)

fn.chi.prop2(u, k = 1.345, deriv = 0, b = 0.71016)
}
\arguments{
\item{u}{numeric vector of evaluation points}

\item{k, c, b}{tunning parameters for each objective function}

\item{deriv}{0 or 1: returns psi(x) or psi'(x)}
}
\description{
Psi functions
}
\details{
see \link{psi.huber}/\link{psi.bisquare} for more details. Note the difference in the returned value for \code{deriv=0} for this package. Functions re-written out of convenience for the development of \link{brlm}.
}
