% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{psi_rho_samp}
\alias{psi_rho_samp}
\alias{fn.loglik.psi_rho}
\alias{fn.proposal.psi_rho}
\alias{fn.logpdf.prop.psirho}
\alias{fn.sample.psi_rho}
\title{Sampling psi_rho: [psi_rho|-]}
\usage{
fn.loglik.psi_rho(psi_rho, mu_rho, rho)

fn.proposal.psi_rho(psi_rhoCur, psi_rho_step)

fn.logpdf.prop.psirho(psi_rho)

fn.sample.psi_rho(psi_rhoCur, mu_rho, rho, psi_rho_step)
}
\arguments{
\item{psi_rho, psi_rhoCur}{desc}

\item{mu_rho}{desc}

\item{rho}{desc}

\item{psi_rho_step}{desc}
}
\description{
Sampling psi_rho: [psi_rho|-]
}
