% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_heirarchical.R
\name{quads}
\alias{quads}
\alias{fn.compute.quad1}
\alias{fn.compute.quad2}
\title{quad1 and qaud2}
\usage{
fn.compute.quad1(Beta, mu0, Sigma0Inv)

fn.compute.quad2(Beta, betalMat, Sigma0Inv)
}
\arguments{
\item{Beta}{Beta}

\item{mu0}{mu0}

\item{betalMat}{matrix of betal's}
}
\description{
quad1 and qaud2
}
\details{
Used for sampling Beta in hierarchical model of paper. quad1 is the current value of (beta-mu0)'Sigma0Inv(beta-mu0). quad2 is the current value of sum_i^nGroups (beta_i-beta) Sigma0Inv (beta_i-beta)
}
