% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_random_effects.R
\name{re_model}
\alias{re_model}
\title{Fit the Normal Theory Random Effects Model}
\usage{
re_model(Y, mu_0, sigma2_0, alpha_t, beta_t, alpha_s, beta_s, nkeep = 1000,
  nburn = 1000)
}
\arguments{
\item{Y}{list; jth element is a vector of the y_ij}

\item{mu_0, sigma2_0}{prior mean and variance for \eqn{\mu}}

\item{alpha_t, beta_t}{prior shape and scale for \eqn{\tau^2}}

\item{alpha_s, beta_s}{prior shape and scale for \eqn{\sigma^2}}

\item{nkeep}{number of iterations to keep}

\item{nburn}{number of iterations to toss}
}
\value{
list with mcmc sample and mean fitted values
}
\description{
Fits the random effects model using standard Gibbs Sampling
\eqn{\mu ~ N(\mu_0, \sigma^2_0)}; \eqn{\tau^2 ~ IG(\alpha_t, \beta_t)}
\eqn{\theta_j~N(\mu, \tau^2), j = 1,2,...,J}; \eqn{\sigma^2~IG(\alpha_s, \beta_s)},
\eqn{y_{ij}~N(\theta_j), \sigma^2), i = 1,2,...,n_j}
}
