% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_random_effects.R
\name{re_sample_mu}
\alias{re_sample_mu}
\title{Sampling Functions for random effects model}
\usage{
re_sample_mu(sigma2_0, J, tau2, mu_0, theta)
}
\description{
Sampling Functions for random effects model
}
