% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_restrictedBayesLm.R
\name{restrictedBayesLm}
\alias{restrictedBayesLm}
\title{MCMC algorithm for Restricted Likelihood Bayesian Linear Model}
\usage{
restrictedBayesLm(
  y,
  X,
  regEst = "Huber",
  scaleEst = "Huber",
  mu0,
  Sigma0,
  a0,
  b0,
  sigma2Int,
  nkeep = 10000,
  nburn = 1000,
  maxit = 400
)
}
\arguments{
\item{y}{vector of repsonse}

\item{X}{design matrix for regression. If intercept is desired, a column of 1's is needed}

\item{regEst}{The estimator of \eqn{\beta}  to condition on. Current options available are \code{Huber} and \code{Tukey} for Huber's and Tukey's (bisquare) estimators, respectively.}

\item{scaleEst}{The estimator of \eqn{\sigma} to condition on. Currently Huber's proposal 2 is the only option available. Specified with \code{Huber}.}

\item{mu0}{prior mean for beta}

\item{Sigma0}{prior var-cov matrix for \eqn{\beta}}

\item{a0}{prior shape parameter for \eqn{\sigma^2}}

\item{b0}{prior scale parameter for \eqn{\sigma^2}}

\item{nkeep}{number of iterations to keep}

\item{nburn}{number of iterations to toss}

\item{maxit}{maximum number of iterations for use in \code{\link[MASS]{rlm}} when computing conditioning statistic during each proposal within the MCMC step for [y | rest, observed statistics]}
}
\value{
list with mcmc sample, mean fitted values, robust regression object from \code{rlm}, and observed conditioning statistics.
}
\description{
Fits the linear regression model
\eqn{\beta~N(\mu_0, \Sigma_0)},
\eqn{\sigma^2~IG(a0, b0)},
\eqn{y~N(X\beta, \sigma^2 I)}
}
\details{
Uses Gibbs sampling to sample from the restricted posterior under the above linear regression model conditioned on estimators \code{regEst} and \code{scaleEst} for \eqn{\beta} and \eqn{\sigma}, respectively. This is just the Gibbs sampler for the full posterior (implemented in \code{brlm::bayesLm} augmented with a step to sample new data given all of the parameters and the observed conditioning statistics
}
