% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{rho_samp}
\alias{rho_samp}
\alias{fn.compute.SigmaRhoInv}
\alias{fn.compute.logDetSigmaRho}
\alias{fn.proposal.rho}
\alias{fn.loglike.rho}
\alias{fn.sample.rho}
\title{sampling for [rho|-] in hierchical model}
\usage{
fn.compute.SigmaRhoInv(rho, K)

fn.compute.logDetSigmaRho(rho, K)

fn.proposal.rho(rhoCur, rho_step)

fn.loglike.rho(rho, Z, alpha_rho, beta_rho)

fn.sample.rho(rhoCur, mu_rho, psi_rho, Z, rho_step)
}
\arguments{
\item{rho}{desc}

\item{K}{desc}

\item{rhoCur}{desc}

\item{rho_step}{desc}

\item{Z}{desc}

\item{alpha_rho}{desc}

\item{beta_rho}{desc}

\item{mu_rho}{desc}

\item{psi_rho}{desc}

\item{Zdesc}{desc}
}
\description{
sampling for [rho|-] in hierchical model
}
\details{
J_nGroups must be defined, I_nGroups must be defined
function to effeciently caclulate Sigma_rhoInv using its special form and a well known matrix inverse result (from the fix-rank kriging chapter 8 of the spatial handbook book); note; the K variable added for the sampling of the z_i's later.
fn.compute.logDetSigmaRho used to effeciently caclulate log(det(Sigma_rho)) using its special form and Sylvester's determinant theorem: det(I_n+AB)=det(I_m+BA)
}
