% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_direct.R
\name{rl_direct}
\alias{rl_direct}
\title{Fitting restricted likelihood location-scale model using direct evaluation.}
\usage{
rl_direct(y, statistic, eta, tau, alpha, beta, mu_lims, sigma2_lims, length_mu,
  length_sigma2, smooth = 1, N)
}
\arguments{
\item{y}{vector of data}

\item{statistic, }{character name of a function computing the location and scale statistic. See details for specification of this function.}

\item{eta, tau}{prior mean and standard deviation for \eqn{mu}}

\item{alpha, beta}{prior shape and scale for \eqn{sigma^2}}

\item{mu_lims, sigma2_lims}{vectors of length 2 defining the limits for the numerical integration necessary to find the normalizing constant in Bayes' Theorem}

\item{length_mu, length_sigma2}{length of the grid in each parameter to do the numerical integration (standard Riemann sum)}

\item{smooth}{scalar or vector of length 2 that will scale the initial bandwidth computed by \code{\link[ks]{hpi}}.}

\item{N}{number of samples of the statistics to use for the kernel density estimation}

\item{...}{additional arguments to be passed to \code{psi}}
}
\value{
A list of length 4: the joint posterior, the two marginals, and the bandwidths used for the kernel density estimate
}
\description{
Full model: \deqn{
\mu~N(\eta, \tau^2),
\sigma^2~IG(\alpha, \beta),
y_1, \dots, y_n~N(\mu, \sigma^2)
}
For the restricted likelihood, conditioning is done on a pair of location and scale statistics \eqn{T(y)=(l(y), s(y))} with the property that \eqn{l(\sigma y+\mu)=\sigma l(y)+\mu} and \eqn{s(\sigma y+\mu)=\sigma s(y)}.
}
\details{
Direct evaluation uses kernel density estimation with a Gaussian kernel to estimate the restricted likelihood. \code{N} specifies the number of samples of the statistics to generate for the kernel density estimate. \code{\link[ks]{hpi}} is used to specify the initial bandwidths independently for the location and the scale. These can be multiplied by \code{smooth} to 'oversmooth' or 'undersmooth' the estimate.

\code{statistic} is a function taking only \code{y} as an input and outputing a two-dimension vector with the location and scale statistic to be conditioned on.

A simple numerical integration using Riemann sums is used to determine the normalizing constant. Parameters for this numerical integration are specified by \code{mu_lims,sigma2_lims, length_mu}, and \code{length_sigma2}.
}
\author{
John R. Lewis \email{lewis.865@osu.edu}
}
