% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_import.R
\name{rl_import}
\alias{rl_import}
\title{Fitting restricted likelihood model using importance sampling technique.}
\usage{
rl_import(y, X, statistic, mu0, Sigma0, alpha, beta, scale, smooth = 1, N,
  Nins)
}
\arguments{
\item{y}{vector of data}

\item{X}{a matrix or data frame containing the explanatory variables. The matrix should include a vector of 1's if intercept is desired.}

\item{statistic}{, character name of a function computing the location and scale statistic. See details for specification of this function.}

\item{alpha}{prior shape and scale for \eqn{sigma^2}}

\item{beta}{prior shape and scale for \eqn{sigma^2}}

\item{scale}{vector length p+1 defining the standard deviations for the normal instrumental distributions for \eqn{\beta} and \eqn{\sigma^2} (with \eqn{\sigma^2} truncated at 0). Importance sample weights should be examined to evaluate the appropriatness of this choice.}

\item{smooth}{scalar or vector of length 2 that will scale the initial bandwidth computed by \code{\link[ks]{hpi}}.}

\item{N}{number of samples of the statistics to use for the kernel density estimation}

\item{Nins}{number of samples from the instrumental distribution}
}
\description{
Full model: \deqn{
\beta~N(\mu_0, \Sigma_0),
\sigma^2~IG(\alpha, \beta),
y~N(X\beta, \sigma^2)
}
For the restricted likelihood, conditioning is done on a pair of location and scale statistics \eqn{T(y)=(b(y), s(y))}. Current implementation allows for these to be a pair of M-estimators as implemented in \code{\link[MASS]{rlm}}
\code{statistic} is a function of \code{X}, \code{y}  and outputing a p+1-dimensional vector with the p location statistics (for \code{beta}) and 1 scale statistic to be conditioned on.
}
\author{
John R. Lewis \email{lewis.865@osu.edu}
}
