% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_import.R
\name{rl_importance}
\alias{rl_importance}
\title{Fitting restricted likelihood model using importance sampling technique.}
\usage{
rl_importance(y, X, statistic, mu0, Sigma0, alpha, beta, cov_b, scale,
  smooth = 1, N, Nins)
}
\arguments{
\item{y}{vector of data}

\item{X}{a matrix or data frame containing the explanatory variables. The matrix should include a vector of 1's if intercept is desired.}

\item{statistic}{character name of a function computing the location and scale statistic. See details for specification of this function.}

\item{alpha}{prior shape and scale for \eqn{sigma^2}}

\item{beta}{prior shape and scale for \eqn{sigma^2}}

\item{cov_b}{positive definite pxp matrix defining the covariance for the normal instrumental distribution on \code{b}. Importance sample weights should be examined to evaluate the appropriatness of this choice.}

\item{scale}{numeric, sd for the truncated normal instrumental distribution for \eqn{\sigma^2}. Importance sample weights should be examined to evaluate the appropriatness of this choice.}

\item{smooth}{scalar or vector of length 2 that will scale the initial bandwidth computed by \code{\link[ks]{hpi}}.}

\item{N}{number of samples of the statistics to use for the kernel density estimation}

\item{Nins}{number of samples from the instrumental distribution}
}
\value{
list with elements \code{impSamps}, \code{w}, \code{fit}. These are the \code{Nins} by \code{length(beta)+1} matrix of importance samples, the corresponding weights, and the observed statistic.
}
\description{
Full model: \deqn{
b~N(\mu_0, \Sigma_0),
\sigma^2~IG(\alpha, \beta),
y~N(Xb, \sigma^2)
}
For the restricted likelihood, conditioning is done on a pair of location and scale statistics \eqn{T(y) = (b(y), s(y))}.
\code{statistic} is a function of \code{X}, \code{y}  and outputing the p+1-dimensional vector \eqn{T(y)} with the p location statistics (for \code{b}) and 1 scale statistic.
Instrumental (importance) distributions are normal. For beta, a multivariate normal centererd at the estimate of beta with covariance \code{cov_b}. For sigma^2, a truncated normal with mean (before truncation) of the estimate of \eqn{\sigma^2} with sd \code{scale}.
}
\author{
John R. Lewis \email{lewis.865@osu.edu}
}
