% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlImportSamp.R
\name{rlImportSamp}
\alias{rlImportSamp}
\title{Fitting restricted likelihood model using importance sampling technique.}
\usage{
rlImportSamp(X, y, psi, scale.est = "Huber", k2 = 1.345, mu0, Sigma0, alpha,
  beta, instDist = NULL, sdInstDist = NULL, smooth = 1, N, Nins,
  maxit = 1000, ...)
}
\arguments{
\item{X}{a matrix or data frame containing the explanatory variables. The matrix should include a vector of 1's if intercept is desired.}

\item{y}{vector of data}

\item{psi}{the \code{psi} function used to define the location estimator. It is a function (possibly given by name) that is described in \code{\link[MASS]{rlm}}. Tuning constants are passed via \code{...}}

\item{scale.est}{specification of the scale estimator (issued by name- either \code{'MAD', 'Huber'}, or \code{'proposal 2'}) and the tuning constant for the Huber proposal 2 scale estimation. See these parameters in \code{\link[MASS]{rlm}} for more information.}

\item{k2}{specification of the scale estimator (issued by name- either \code{'MAD', 'Huber'}, or \code{'proposal 2'}) and the tuning constant for the Huber proposal 2 scale estimation. See these parameters in \code{\link[MASS]{rlm}} for more information.}

\item{alpha}{prior shape and scale for \eqn{sigma^2}}

\item{beta}{prior shape and scale for \eqn{sigma^2}}

\item{instDist}{Placeholder to allow for more user specific instrumental distributions. Currently not used.}

\item{sdInstDist}{Vector of length 2 defining the scale for the instrumental distribution when instDist=NULL. In this case, the instrumental distribution for \eqn{\mu} and \eqn{\sigma^2} are independent normal and log normal distributions, respectively. The first value is then the standard deviation of the normal. The second is the standard deviation of \eqn{\log\sigma^2} (i.e. the \code{sdlog} argument in \code{\link[stats]{Lognormal}}). If left as NULL a mutltiple of 5 is used on the asymptotic variance covariance matrix. Importance sample weights should be examined to evaluate the appropriatness of this choice.}

\item{smooth}{scalar or vector of length 2 that will scale the initial bandwidth computed by \code{\link[ks]{hpi}}.}

\item{N}{number of samples of the statistics to use for the kernel density estimation}

\item{Nins}{number of samples from the instrumental distribution}

\item{maxit}{the limit on the number of IWLS iterations. Same as in \code{\link[MASS]{rlm}}}

\item{...}{arguments to pass to the psi functions}
}
\description{
Full model: \deqn{
\beta~N(\mu_0, \Sigma_0),
\sigma^2~IG(\alpha, \beta),
y~N(X\beta, \sigma^2)
}
For the restricted likelihood, conditioning is done on a pair of location and scale statistics \eqn{T(y)=(b(y), s(y))}. Current implementation allows for these to be a pair of M-estimators as implemented in \code{\link[MASS]{rlm}}
}
\details{
Fill in some details
}
\examples{
# Example using the Belgium Phone calls data
# Conditioning statistics are simultaneous M-estimators of coefficients and scale.
In the example below, Tukeys bisquare is used for the coefficients and median absolute deviation (MAD) is used for the scale. Note the sample sizes are dramatically reduces to run quickly.

nForPrior<-3 #number of data points to use formulate prior
data(MASS::phones) #load a prepare data
phones<-data.frame(cbind(phones$year, phones$calls))
names(phones)<-c("year", "calls")
phones$year<-phones$year-mean(phones$year) #center
phonesFit<-phones[-c(1:nForPrior),]
n<-nrow(phonesFit)
X<-cbind(rep(1, n), phonesFit$year)
y<-phonesFit$calls

# formulate priors
phonesPrior<-phones[c(1:nForPrior),]
priorFit<-lm(calls~year, data=phonesPrior)
summary(priorFit)
mu0<-coef(priorFit)
sigma2Hat<-summary(priorFit)$sigma^2
alpha<-2
beta<-(alpha-1) #sigma2Hat*(alpha-1)

# choosing how much to inlfate fitted variance covariance matrix from the prior data fit
inflate<-10
Sigma0<-vinflate[inflate]*vcov(priorFit)

 # Fit restricted likelihood model using importance sampling
restFit<-rlImportSamp(X=X,y=y, psi=psi.bisquare, scale.est='MAD', mu0=mu0, Sigma0=Sigma0, alpha=alpha, beta=beta, smooth=1,N=1e1,Nins=1e1, maxit=10000)

# Some plots
plot(tst$w, cex=.1, pch=19)
plot(density(tst$impSamps[,1],weights=tst$w))
abline(v=coef(restFit$fit)[1], col=2)
plot(density(tst$impSamps[,2],weights=tst$w))
abline(v=coef(restFit$fit)[2], col=2)
plot(density(tst$impSamps[,3],weights=tst$w))
abline(v=summary(restFit$fit)$sigma, col=2)
}
\author{
John R. Lewis \email{lewis.865@osu.edu}
}

