% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hierarchical.R
\name{z_l}
\alias{z_l}
\alias{fn.compute.sigma2}
\alias{fn.compute.Z2}
\alias{fn.compute.Z}
\alias{fn.hat}
\alias{fn.logLike.zl}
\alias{fn.proposal.zl}
\alias{fn.logpdf.proposal.Z}
\alias{fn.sample.Zl}
\alias{fn.one.rep.Z}
\title{[Zl|-]: sample one at a time}
\usage{
fn.compute.sigma2(Z, a0, b0)

fn.compute.Z2(sigma2, a0, b0)

fn.compute.Z(sigma2, a0, b0)

fn.hat(Xl, betal)

fn.logLike.zl(zl, zminus, yl, fitsl, condsd, condMeanMult)

fn.proposal.zl(zlCur, step = 1)

fn.logpdf.proposal.Z(zlProp, zlCur, step)

fn.sample.Zl(zlCur, zminus, yl, fitsl, condsd, condMeanMult, step = 1)

fn.one.rep.Z(Zcur, y, X, betalMat, rho, step = rep(1, length(Zcur)))
}
\arguments{
\item{Z, sigma2}{desc}

\item{a0, b0}{desc}

\item{Xl, bl}{desc}

\item{zl, zlCur, zlProp}{desc}

\item{zminus, }{desc}

\item{yl, }{desc}

\item{fitsl, }{desc}

\item{condsd, }{desc}

\item{condMeanMult}{desc}

\item{step}{desc}
}
\description{
[Zl|-]: sample one at a time
}
\details{
function to compute simga2_is from a z vector (or a single z) from nGroups dimensional Normal distribution with marginal means 0 and vars 1
}
